ImageModule/requirements.txt
openai>=1.0.0
google-genai>=0.3.0
pillow>=9.0.0
rembg>=2.0.0
requests>=2.28.0
python-dotenv>=0.19.0
cloudinary>=1.30.0
beautifulsoup4>=4.11.0


pathlib>=1.0.1